Added implementation for checkout endpoint and changed some values in database to match examples from provided file. Need to write tests now to test not caught cases.

+++ src/main/java/org/js/checkoutcomponent/model/CheckoutResponse.java
@@ -15,5 +15,5 @@ public class CheckoutResponse {
     @Schema(name = "itemPrices", description = "Detailed breakdown of prices per item")
     private List<ItemPrice> itemPrices;
     @Schema(name = "bundleDiscountApplied", example = "5.0", description = "Total bundle discounts applied")
-    private double bundleDiscountApplied;
+    private BigDecimal bundleDiscountApplied;
 }
